+++ src/main/java/a4/KU_TY_backend/KU_TY_backend/entity/Event.java
@@ -43,6 +43,10 @@ public class Event {
     @ManyToOne
     @JoinColumn(name = "type_id")
     private EventType eventType;
+    @OneToMany(mappedBy = "event",cascade = CascadeType.ALL)
+    @JsonIgnore
+    private List<Feedback> feedbackList;
+
     public EventResponse toResponse() {
         EventResponse dto = new EventResponse();
         dto.setEventId(eventId);
