+++ src/main/java/a4/KU_TY_backend/KU_TY_backend/entity/Feedback.java
@@ -4,20 +4,21 @@ package a4.KU_TY_backend.KU_TY_backend.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.util.UUID;
+
 @Entity
 @Data
 @Table(name = "feedback")
 public class Feedback {
-    @EmbeddedId
-    private FeedbackKey key;
+    @Id
+    @GeneratedValue
+    private UUID feedbackId;
     @ManyToOne
-    @MapsId("userId")
-    @JoinColumn(name = "user_id")
-    private User user;
-    @ManyToOne
-    @MapsId("eventId")
-    @JoinColumn(name = "event_id")
+    @JoinColumn(name = "event_id", nullable = false)
     private Event event;
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
     @Column(nullable = false)
     private String feedback;
