--- src/main/java/a4/KU_TY_backend/KU_TY_backend/entity/FeedbackKey.java
@@ -1,24 +1,0 @@
-package a4.KU_TY_backend.KU_TY_backend.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-@Embeddable
-public class FeedbackKey implements Serializable {
-    @Column(name = "user_id")
-    private UUID userId;
-    @Column(name = "event_id")
-    private UUID eventId;
-
-}
