+++ src/main/java/a4/KU_TY_backend/KU_TY_backend/entity/User.java
@@ -39,6 +39,9 @@ public class User {
     @JsonIgnore
     private List<EventUser> joinedEventList;
     private String imageUrl;
+    @OneToMany(mappedBy = "user")
+    @JsonIgnore
+    private List<Feedback> givedFeedbackList;
 
     public UserResponse toResponse(){
         UserResponse response = new UserResponse();
