+++ src/main/java/a4/KU_TY_backend/KU_TY_backend/repository/FeedbackRepository.java
@@ -1,10 +1,11 @@
 package a4.KU_TY_backend.KU_TY_backend.repository;
 
-import a4.KU_TY_backend.KU_TY_backend.entity.EventUser;
-import a4.KU_TY_backend.KU_TY_backend.entity.EventUserKey;
 import a4.KU_TY_backend.KU_TY_backend.entity.Feedback;
-import a4.KU_TY_backend.KU_TY_backend.entity.FeedbackKey;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface FeedbackRepository  extends JpaRepository<Feedback, FeedbackKey> {
+import java.util.UUID;
+
+@Repository
+public interface FeedbackRepository  extends JpaRepository<Feedback, UUID> {
 }
